SciChart.iOS v3.0.0.4810

+++ Examples/SciChartDemo/SciChartDemo/Examples/Examples2D/StylingAndTheming/UsePaletteProvider/OhlcCustomPaletteProvider.m
@@ -15,6 +15,7 @@
 //******************************************************************************
 
 #import "OhlcCustomPaletteProvider.h"
+#import <SciChart/SCIPaletteProviderBase+Protected.h>
 
 @implementation OhlcCustomPaletteProvider {
     SCIIntegerValues *_colors;
@@ -38,7 +39,7 @@
     SCIOhlcRenderPassData *rpd = (SCIOhlcRenderPassData *)rSeries.currentRenderPassData;
     SCIDoubleValues *xValues = rpd.xValues;
     
-    NSInteger count = rpd.pointsCount;
+    int count = rpd.pointsCount;
     _colors.count = count;
     
     double x1 = _annotation.x1.toDouble;
