+++ Examples/SciChartDemo/SciChartDemo/Examples/Examples2D/StylingAndTheming/UsePaletteProvider/XyCustomPaletteProvider.m
@@ -15,6 +15,7 @@
 //******************************************************************************
 
 #import "XyCustomPaletteProvider.h"
+#import <SciChart/SCIPaletteProviderBase+Protected.h>
 
 @implementation XyCustomPaletteProvider {
     SCIUnsignedIntegerValues *_colors;
@@ -37,7 +38,7 @@
     SCIXyRenderPassData *rpd = (SCIXyRenderPassData *)rSeries.currentRenderPassData;
     SCIDoubleValues *xValues = rpd.xValues;
     
-    NSInteger count = rpd.pointsCount;
+    int count = rpd.pointsCount;
     _colors.count = count;
     
     double x1 = _annotation.x1.toDouble;
